Extract distributor library
